large restructuring of the package internal structure and tests
