+++ sentinelhub/__init__.py
@@ -3,6 +3,28 @@ This module lists all externally useful classes and functions
 """
 
 from ._version import __version__
+from .api import (
+    BatchCollection,
+    BatchRequest,
+    BatchRequestStatus,
+    BatchTileStatus,
+    BatchUserAction,
+    ByocCollection,
+    ByocCollectionAdditionalData,
+    ByocTile,
+    FisRequest,
+    SentinelHubBatch,
+    SentinelHubBYOC,
+    SentinelHubCatalog,
+    SentinelHubRequest,
+    SentinelHubStatistical,
+    WcsRequest,
+    WebFeatureService,
+    WmsRequest,
+    monitor_batch_job,
+    opensearch,
+)
+from .api.opensearch import get_area_dates, get_area_info, get_tile_info, get_tile_info_id
 from .areas import (
     BatchSplitter,
     BBoxSplitter,
@@ -16,8 +38,13 @@ from .config import SHConfig
 from .constants import CRS, CustomUrlParam, HistogramType, MimeType, ServiceType, ServiceUrl, SHConstants
 from .data_collections import DataCollection, DataSource
 from .data_collections_bands import Band, Unit
-from .data_request import FisRequest, WcsRequest, WmsRequest
-from .download import DownloadClient, DownloadRequest, SentinelHubDownloadClient, SentinelHubStatisticalDownloadClient
+from .download import (
+    DownloadClient,
+    DownloadRequest,
+    SentinelHubDownloadClient,
+    SentinelHubSession,
+    SentinelHubStatisticalDownloadClient,
+)
 from .exceptions import AwsDownloadFailedException, DownloadFailedException
 from .geo_utils import (
     bbox_to_dimensions,
@@ -36,21 +63,5 @@ from .geo_utils import (
 from .geometry import BBox, BBoxCollection, Geometry
 from .geopedia import GeopediaFeatureIterator, GeopediaImageRequest, GeopediaSession, GeopediaWmsRequest
 from .io_utils import read_data, write_data
-from .ogc import WebFeatureService
-from .opensearch import get_area_dates, get_area_info, get_tile_info, get_tile_info_id
 from .os_utils import create_parent_folder, get_content_list, get_file_list, get_folder_list, make_folder, rename, size
-from .sentinelhub_batch import (
-    BatchCollection,
-    BatchRequest,
-    BatchRequestStatus,
-    BatchTileStatus,
-    BatchUserAction,
-    SentinelHubBatch,
-    monitor_batch_job,
-)
-from .sentinelhub_byoc import ByocCollection, ByocCollectionAdditionalData, ByocTile, SentinelHubBYOC
-from .sentinelhub_catalog import SentinelHubCatalog
-from .sentinelhub_request import SentinelHubRequest
-from .sentinelhub_session import SentinelHubSession
-from .sentinelhub_statistical import SentinelHubStatistical
 from .time_utils import filter_times, is_valid_time, parse_time, parse_time_interval, serialize_time
