+++ sentinelhub/api/__init__.py
@@ -1,0 +1,20 @@
+"""
+The part of the package that implements interface with Sentinel Hub services.
+"""
+
+from .batch import (
+    BatchCollection,
+    BatchRequest,
+    BatchRequestStatus,
+    BatchTileStatus,
+    BatchUserAction,
+    SentinelHubBatch,
+    monitor_batch_job,
+)
+from .byoc import ByocCollection, ByocCollectionAdditionalData, ByocTile, SentinelHubBYOC
+from .catalog import SentinelHubCatalog
+from .fis import FisRequest
+from .ogc import WcsRequest, WmsRequest
+from .process import SentinelHubRequest
+from .statistical import SentinelHubStatistical
+from .wfs import WebFeatureService
