+++ sentinelhub/download/__init__.py
@@ -6,3 +6,4 @@ from .client import DownloadClient
 from .request import DownloadRequest
 from .sentinelhub_client import SentinelHubDownloadClient
 from .sentinelhub_statistical_client import SentinelHubStatisticalDownloadClient
+from .session import SentinelHubSession
