+++ tests/api/test_ogc.py
@@ -5,7 +5,6 @@ from typing import List, Optional, Type, Union
 import numpy as np
 import pytest
 from numpy.testing import assert_array_equal
-from shapely.geometry import MultiPolygon
 
 from sentinelhub import (
     CRS,
@@ -16,11 +15,9 @@ from sentinelhub import (
     MimeType,
     ServiceType,
     WcsRequest,
-    WebFeatureService,
     WmsRequest,
 )
-from sentinelhub.data_request import OgcRequest
-from sentinelhub.ogc import OgcImageService
+from sentinelhub.api.ogc import OgcImageService, OgcRequest
 from sentinelhub.testing_utils import get_output_folder, test_numpy_data
 
 pytestmark = pytest.mark.sh_integration
@@ -536,39 +533,3 @@ def test_too_large_request():
 
     with pytest.raises(DownloadFailedException):
         request.get_data()
-
-
-@pytest.mark.parametrize(
-    "args, kwargs, expected_len",
-    [
-        (
-            [
-                BBox(bbox=(-5.23, 48.0, -5.03, 48.17), crs=CRS.WGS84),
-                (datetime.date(year=2017, month=1, day=5), datetime.date(year=2017, month=12, day=16)),
-            ],
-            dict(data_collection=DataCollection.SENTINEL2_L1C, maxcc=0.1),
-            13,
-        ),
-        (
-            [BBox(bbox=(-5.23, 48.0, -5.03, 48.17), crs=CRS.WGS84), "latest"],
-            dict(data_collection=DataCollection.SENTINEL2_L2A),
-            1,
-        ),
-    ],
-)
-def test_wfs(args, kwargs, expected_len):
-    iterator = WebFeatureService(*args, **kwargs)
-    features = list(iterator)
-    dates = iterator.get_dates()
-    geometries = iterator.get_geometries()
-    tiles = iterator.get_tiles()
-
-    for result_list, expected_type in [
-        (features, dict),
-        (dates, datetime.datetime),
-        (geometries, MultiPolygon),
-        (tiles, tuple),
-    ]:
-        assert len(result_list) == expected_len
-        for result in result_list:
-            assert isinstance(result, expected_type)
