+++ tests/api/test_opensearch.py
@@ -1,3 +1,6 @@
+"""
+Tests for Sentinel Hub Opensearch service interface
+"""
 from sentinelhub import CRS, BBox, get_area_dates, get_tile_info
