+++ tests/api/test_sentinelhub_batch.py
