+++ tests/api/test_sentinelhub_statistical.py
